change for_new_model

+++ src/predict_app.py
@@ -6,7 +6,7 @@ from flask_cors import CORS
 from joblib import load
 from flask_httpauth import HTTPTokenAuth
 from flask import send_from_directory
-from src.utils import predict_io_bounded, predict_cpu_bounded, predict_cpu_multithread
+from utils import predict_io_bounded, predict_cpu_bounded, predict_cpu_multithread
 
 MODEL_SAVE_PATH = 'models/linear_regression_v01.joblib'
 
@@ -37,7 +37,7 @@ def predict(in_data: dict) -> int:
     :rtype: int
     """
     area = float(in_data['total_meters'])
-    rooms_count = float(in_data['rooms_count'])
+    rooms_count = int(in_data['rooms_count'])
     floor = int(in_data['floor'])
     floors_count = int(in_data['floors_count'])
     is_first = (floor == 1)
